♻️ refactor(svelte): use component based structure
