+++ src/components/header.svelte
@@ -1,0 +1,10 @@
+<h1
+	class="w-full py-6 text-center text-2xl font-extrabold text-slate-900 dark:text-gray-300 md:text-4xl"
+>
+	F1 2022 AI Difficulty Calculator
+</h1>
+<h2 class="text-justify text-lg text-gray-700 dark:text-gray-300 sm:text-base">
+	<span class="font-extrabold">Usage: </span>Input your best lap time from
+	<span class="font-extrabold underline">Time Trial Equal Performance Mode</span>
+	to see your recommended AI level
+</h2>
