+++ src/components/trackSelect.svelte
@@ -1,0 +1,22 @@
+<script lang="ts">
+	import type { Track } from 'src/app/TrackUtils';
+	export let selected: string;
+	export let tracks: { [key: string]: Track };
+</script>
+
+<div class="flex w-full flex-col md:w-3/4 xl:w-1/2">
+	<label for="trackSelect" class="text-center text-gray-700 dark:text-gray-300 sm:text-left"
+		>Select Track</label
+	>
+	<select
+		id="trackSelect"
+		bind:value={selected}
+		class="rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-gray-900 focus:border-blue-500 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:text-gray-300 dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+	>
+		{#each Object.entries(tracks) as [track, trackInfo]}
+			<option value={track}
+				>{trackInfo.flagmoji + trackInfo.friendlyName + ' - ' + trackInfo.trackName}</option
+			>
+		{/each}
+	</select>
+</div>
